AirflowModel and python airflow computation added

See merge request vadere/vadere!273

+++ VadereSimulator/src/org/vadere/simulator/projects/migration/jsontranformation/json/TargetVersionV2_1.java
@@ -347,7 +347,7 @@ public class TargetVersionV2_1 extends SimpleJsonTransformation {
         AttributesAirTransmissionModelAerosolCloud aerosolCloudParameters = new AttributesAirTransmissionModelAerosolCloud(
                 tempData.get("aerosolCloudHalfLife"),
                 initialRadius,
-                Math.pow(10, tempData.get("pedestrianPathogenEmissionCapacity")),
+                (int) Math.pow(10, tempData.get("pedestrianPathogenEmissionCapacity")),
                 0,
                 0.0125,
                 tempData.get("pedestrianPathogenAbsorptionRate")
